DSTA-SLR/requirements-conda.txt
numpy==2.4.3
scipy==1.17.1
tqdm==4.67.3
torchmetrics==1.9.0
thop==0.1.1.post2209072238
wandb==0.25.1
timm==1.0.25
einops==0.8.2
PyYAML==6.0.3
pillow==12.1.1
requests==2.32.5
GitPython==3.1.46
mega.py==1.0.8
pypdf==6.8.0
